@@ -3,7 +3,7 @@
 	"description": "A handy tool for comparing structured data quickly in a key-value manner",
 	"license": "MIT",
 	"require": {
-		"php": ">= 7.1",
+		"php": ">= 8.1",
 
 		"ext-pdo": "*",
 		"ext-pdo_sqlite": "*",
@@ -11,7 +11,7 @@
 		"ext-mbstring": "*"
 	},
 	"require-dev": {
-		"phpunit/phpunit": ">= 7.0, < 10.0",
+		"phpunit/phpunit": "^10.0",
 		"phpstan/phpstan": ">= 0.1"
 	},
 	"suggest": {
