@@ -1,6 +1,11 @@
 <?php
 namespace DataDiff;
 
+/**
+ * @template TKeySpec of array<string, mixed>
+ * @template TValueSpec of array<string, mixed>
+ * @template TExtraSpec of array<string, mixed>
+ */
 interface DiffStorageInterface {
 	/**
 	 * @return string[]
@@ -8,12 +13,12 @@ interface DiffStorageInterface {
 	public function getKeys(): array;
 
 	/**
-	 * @return DiffStorageStore
+	 * @return DiffStorageStore<TKeySpec, TValueSpec&TExtraSpec, TKeySpec&TValueSpec&TExtraSpec>
 	 */
 	public function storeA(): DiffStorageStore;
 
 	/**
-	 * @return DiffStorageStore
+	 * @return DiffStorageStore<TKeySpec, TValueSpec&TExtraSpec, TKeySpec&TValueSpec&TExtraSpec>
 	 */
 	public function storeB(): DiffStorageStore;
 }
